Janus: Extract publisher
